requirements.txt
fastapi
strawberry-graphql[fastapi]
neo4j
python-dotenv
uvicorn